--- crates/gen/.gitattributes
@@ -1,3 +1,0 @@
-*       text=auto
-*       text eol=lf
-*.png   -text
