--- crates/gen/.gitignore
@@ -1,4 +1,0 @@
-target
-Cargo.lock
-.vscode
-.vs
