+++ crates/gen/Cargo.toml
@@ -13,5 +13,4 @@ syn = "1.0"
 proc-macro2 = "1.0"
 squote = "0.1.2"
 sha1 = "0.6.0"
-serde_json = "1.0"
 rayon = "1.3"
