+++ crates/gen/src/lib.rs
@@ -7,12 +7,10 @@ mod type_tree;
 mod types;
 mod winmd;
 
-pub mod dependencies;
 use format_ident::format_ident;
 use to_snake::to_snake;
 pub use type_limits::{NamespaceTypes, TypeLimit, TypeLimits};
 pub use type_reader::TypeReader;
 pub use type_tree::TypeTree;
 pub use types::*;
-pub use winmd::AttributeArg;
 use winmd::*;
