+++ crates/macros/Cargo.toml
@@ -18,3 +18,4 @@ cargo_toml = "0.8"
 winrt-gen = { path = "../gen", version = "0.7.2" }
 rayon = "1.3"
 winrt-deps = { path = "../deps", version = "0.1" }
+serde_json = "1.0"
